feat(mobile): add refetch indicator to  swap review
